+++ kroxylicious-runtime/src/main/java/io/kroxylicious/proxy/internal/filter/NettyFilterContext.java
@@ -6,6 +6,8 @@
 
 package io.kroxylicious.proxy.internal.filter;
 
+import java.util.Set;
+
 import io.netty.channel.EventLoop;
 
 import io.kroxylicious.proxy.config.PluginFactory;
@@ -30,9 +32,18 @@ public class NettyFilterContext implements FilterFactoryContext {
     }
 
     @Override
-    public <P> P pluginInstance(Class<P> pluginClass, String instanceName) {
-        PluginFactory<P> pluginFactory = pluginFactoryRegistry.pluginFactory(pluginClass);
-        return pluginFactory.pluginInstance(instanceName);
+    public <P> P pluginInstance(Class<P> pluginClass, String implementationName) {
+        PluginFactory<P> pluginFactory = pluginFactory(pluginClass);
+        return pluginFactory.pluginInstance(implementationName);
+    }
+
+    @Override
+    public <P> Set<String> pluginImplementationNames(Class<P> pluginClass) {
+        return pluginFactory(pluginClass).registeredInstanceNames();
+    }
+
+    private <P> PluginFactory<P> pluginFactory(Class<P> pluginClass) {
+        return pluginFactoryRegistry.pluginFactory(pluginClass);
     }
 
 }